utils(store): split up various stores into their own modules
