+++ compiler/hash-utils/src/store/sequence.rs
@@ -1,8 +1,7 @@
-//! Provides generic data structures to store values by generated keys in an
-//! efficient way, with interior mutability.
+//! Defines a [SequenceStore] and [SequenceStoreKey] for a given type. Such a
+//! store keeps a sequence of elements per entry, this is useful for interning
+//! a list of items as a single id.
 
-// @@Organisation: Move this module to the `hash_alloc` crate and split it into
-// smaller modules.
 use std::{
     hash::Hash,
     iter::{repeat, Map, Repeat, Zip},
@@ -10,219 +9,10 @@ use std::{
     ops::Range,
 };
 
-use dashmap::{
-    mapref::one::{Ref, RefMut},
-    DashMap,
-};
-use fxhash::FxBuildHasher;
 pub use fxhash::{FxHashMap, FxHashSet};
 use parking_lot::{
     MappedRwLockReadGuard, MappedRwLockWriteGuard, RwLock, RwLockReadGuard, RwLockWriteGuard,
 };
-
-/// Represents a key that can be used to index a [`Store`].
-pub trait StoreKey: Copy + Eq {
-    /// Turn the key into an index.
-    fn to_index(self) -> usize;
-    /// Create a key from an index.
-    ///
-    /// This should generally not be used in client code.
-    fn from_index_unchecked(index: usize) -> Self;
-}
-
-/// Create a new [`StoreKey`] with the given name.
-#[macro_export]
-macro_rules! new_store_key {
-    ($visibility:tt $name:ident $(, derives = $($extra_derives:ident),*)?) => {
-        #[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Hash, $($($extra_derives),*)?)]
-        $visibility struct $name {
-            index: u32,
-        }
-
-        impl $crate::store::StoreKey for $name {
-
-            fn to_index(self) -> usize {
-                self.index.try_into().unwrap()
-            }
-
-            fn from_index_unchecked(index: usize) -> Self {
-                Self { index: index.try_into().unwrap() }
-            }
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! impl_debug_for_store_key {
-    ($name:ident) => {
-        impl std::fmt::Debug for $name {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                f.debug_tuple(stringify!($name)).field(&self.index).finish()
-            }
-        }
-    };
-}
-
-/// The internal data of a store.
-pub type StoreInternalData<Value> = RwLock<Vec<Value>>;
-pub type StoreBorrowHandle<'a, Value> = MappedRwLockReadGuard<'a, Value>;
-pub type StoreBorrowMutHandle<'a, Value> = MappedRwLockWriteGuard<'a, Value>;
-
-/// A store, which provides a way to efficiently store values indexed by opaque
-/// generated keys.
-///
-/// Use [`new_store_key`] to make such a key type, and [`new_store`] to make a
-/// store.
-///
-/// This trait operates on interior mutability and so all of its methods take
-/// `&self`. This makes it easy to use from within contexts without having to
-/// worry too much about borrowing rules.
-///
-/// *Warning*: The `Value`'s `Clone` implementation must not interact with the
-/// store, otherwise it might lead to a panic.
-pub trait Store<Key: StoreKey, Value> {
-    /// Get a reference to the internal data of the store.
-    ///
-    /// This should only be used to implement new store methods, not to access
-    /// the store.
-    fn internal_data(&self) -> &StoreInternalData<Value>;
-
-    /// Create a value inside the store, given its key, returning its key.
-    fn create_with(&self, value_fn: impl FnOnce(Key) -> Value) -> Key {
-        let next_index = self.internal_data().read().len();
-        let key = Key::from_index_unchecked(next_index);
-        let value = value_fn(key);
-        let mut data = self.internal_data().write();
-        data.push(value);
-        key
-    }
-
-    /// Create a value inside the store, returning its key.
-    fn create(&self, value: Value) -> Key {
-        let mut data = self.internal_data().write();
-        let next_index = data.len();
-        data.push(value);
-        Key::from_index_unchecked(next_index)
-    }
-
-    /// Get a value by a key, and map it to another value given its reference.
-    ///
-    /// *Warning*: Do not call mutating store methods (`create` etc) in `f`
-    /// otherwise there will be a panic. If you want to do this, consider using
-    /// [`CloneStore::map()`] instead.
-    fn map_fast<T>(&self, key: Key, f: impl FnOnce(&Value) -> T) -> T {
-        let data = self.internal_data().read();
-        let value = data.get(key.to_index()).unwrap();
-        f(value)
-    }
-
-    /// Get many values by a collection of keys, and map them to another value.
-    ///
-    /// *Warning*: Do not call mutating store methods (`create` etc) in `f`
-    /// otherwise there will be a panic. If you want to do this, consider using
-    /// [`CloneStore::map_many()`] instead.
-    fn map_many_fast<T>(
-        &self,
-        keys: impl IntoIterator<Item = Key>,
-        f: impl FnOnce(&[&Value]) -> T,
-    ) -> T {
-        let data = self.internal_data().read();
-        let values =
-            keys.into_iter().map(|key| data.get(key.to_index()).unwrap()).collect::<Vec<_>>();
-        f(&values)
-    }
-
-    /// Borrow a value mutably, given a key. Returns an appropriate handle which
-    /// implements `DerefMut` to the value.
-    fn borrow_mut(&self, key: Key) -> StoreBorrowMutHandle<'_, Value> {
-        let data = self.internal_data().write();
-        RwLockWriteGuard::map(data, |d| d.get_mut(key.to_index()).unwrap())
-    }
-
-    /// Borrow a value, given a key. Returns an appropriate handle which
-    /// implements `Deref` to the value.
-    fn borrow(&self, key: Key) -> StoreBorrowHandle<'_, Value> {
-        let data = self.internal_data().read();
-        RwLockReadGuard::map(data, |d| d.get(key.to_index()).unwrap())
-    }
-
-    /// Modify a value by a key, possibly returning another value.
-    ///
-    /// *Warning*: Do not call mutating store methods (`create` etc) in `f`
-    /// otherwise there will be a panic. If you want to do this, consider using
-    /// [`CloneStore::modify()`] instead.
-    fn modify_fast<T>(&self, key: Key, f: impl FnOnce(&mut Value) -> T) -> T {
-        let mut data = self.internal_data().write();
-        let value = data.get_mut(key.to_index()).unwrap();
-        f(value)
-    }
-}
-
-/// Additional functionality for [`Store`] when the value implements [`Clone`].
-pub trait CloneStore<Key: StoreKey, Value: Clone>: Store<Key, Value> {
-    /// Get a value by its key.
-    fn get(&self, key: Key) -> Value {
-        self.internal_data().read().get(key.to_index()).unwrap().clone()
-    }
-
-    /// Get a value by a key, and map it to another value given its reference.
-    ///
-    /// It is safe to provide a closure `f` to this function that modifies the
-    /// store in some way (`create` etc). If you do not need to modify the
-    /// store, consider using [`Store::map_fast()`] instead.
-    fn map<T>(&self, key: Key, f: impl FnOnce(&Value) -> T) -> T {
-        let value = self.get(key);
-        f(&value)
-    }
-
-    /// Modify a value by a key, possibly returning another value.
-    ///
-    /// It is safe to provide a closure `f` to this function that modifies the
-    /// store in some way (`create` etc). If you do not need to modify the
-    /// store, consider using [`Store::modify_fast()`] instead.
-    fn modify<T>(&self, key: Key, f: impl FnOnce(&mut Value) -> T) -> T {
-        let mut value = self.get(key);
-        let ret = f(&mut value);
-        self.set(key, value);
-        ret
-    }
-
-    /// Set a key's value to a new value, returning the old value.
-    fn set(&self, key: Key, new_value: Value) -> Value {
-        let mut data = self.internal_data().write();
-        let value_ref = data.get_mut(key.to_index()).unwrap();
-        let old_value = value_ref.clone();
-        *value_ref = new_value;
-        old_value
-    }
-}
-
-impl<Key: StoreKey, Value: Clone, S: Store<Key, Value>> CloneStore<Key, Value> for S {}
-
-/// A default implementation of [`Store`].
-#[derive(Debug)]
-pub struct DefaultStore<K, V> {
-    data: StoreInternalData<V>,
-    _phantom: PhantomData<K>,
-}
-
-impl<K, V> std::default::Default for DefaultStore<K, V> {
-    fn default() -> Self {
-        Self { data: RwLock::new(Vec::new()), _phantom: PhantomData }
-    }
-}
-
-impl<K, V> DefaultStore<K, V> {
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
-
-impl<K: StoreKey, V: Clone> Store<K, V> for DefaultStore<K, V> {
-    fn internal_data(&self) -> &StoreInternalData<V> {
-        &self.data
-    }
-}
 
 #[derive(Copy, Clone, Default, PartialEq, Eq, Hash)]
 pub struct SequenceStoreElement<T>(pub T, pub usize);
@@ -379,7 +169,7 @@ macro_rules! new_sequence_store_key_direct {
             }
         }
 
-        impl $crate::store::SequenceStoreKey for $name {
+        impl $crate::store::sequence::SequenceStoreKey for $name {
             type ElementKey = $el_name;
 
             fn to_index_and_len(self) -> (usize, usize) {
@@ -780,169 +570,11 @@ impl<K: SequenceStoreKey, V: Clone> SequenceStore<K, V> for DefaultSequenceStore
     }
 }
 
-/// The internal data structure for a [`PartialStore`].
-pub type PartialStoreInternalData<Key, Value> = DashMap<Key, Value, FxBuildHasher>;
-pub type PartialStoreBorrowHandle<'a, Key, Value> = Ref<'a, Key, Value, FxBuildHasher>;
-pub type PartialStoreBorrowMutHandle<'a, Key, Value> = RefMut<'a, Key, Value, FxBuildHasher>;
-
-/// A partial store, which provides a way to store values indexed by existing
-/// keys. Unlike [`Store`], not every instance of `Key` necessarily has a value
-/// in a [`PartialStore<Key, _>`].
-///
-/// Use [`new_partial_store`] to make a new partial store.
-///
-/// Like [`Store`], this data structure has interior mutability and so all of
-/// its methods take `&self`. This makes it easy to use from within contexts
-/// without having to worry too much about borrowing rules.
-///
-/// *Warning*: The `Value`'s `Clone` implementation must not interact with the
-/// store, otherwise it might lead to a panic.
-pub trait PartialStore<Key: Copy + Eq + Hash, Value> {
-    fn internal_data(&self) -> &PartialStoreInternalData<Key, Value>;
-
-    /// Insert a key-value pair inside the store, returning the old value if it
-    /// exists.
-    fn insert(&self, key: Key, value: Value) -> Option<Value> {
-        self.internal_data().insert(key, value)
-    }
-
-    /// Whether the store has the given key.
-    fn has(&self, key: Key) -> bool {
-        self.internal_data().contains_key(&key)
-    }
-
-    /// Get a value by a key, and map it to another value given its reference,
-    /// if it exists.
-    ///
-    /// *Warning*: Do not call mutating store methods (`create` etc) in `f`
-    /// otherwise there will be a panic. If you want to do this, consider using
-    /// [`Self::map()`] instead.
-    fn map_fast<T>(&self, key: Key, f: impl FnOnce(Option<&Value>) -> T) -> T {
-        let data = self.internal_data();
-        let value = data.get(&key);
-        match value {
-            Some(value) => f(Some(value.value())),
-            None => f(None),
-        }
-    }
-
-    /// Modify a value by a key, possibly returning another value, if it exists.
-    ///
-    /// *Warning*: Do not call mutating store methods (`create` etc) in `f`
-    /// otherwise there will be a panic. If you want to do this, consider using
-    /// [`Self::modify()`] instead.
-    fn modify_fast<T>(&self, key: Key, f: impl FnOnce(Option<&mut Value>) -> T) -> T {
-        let data = self.internal_data();
-        let value = data.get_mut(&key);
-        match value {
-            Some(mut value) => f(Some(value.value_mut())),
-            None => f(None),
-        }
-    }
-
-    /// Borrow a value mutably, given a key. Returns an appropriate handle which
-    /// implements `DerefMut` to the value.
-    fn borrow_mut(&self, key: Key) -> PartialStoreBorrowMutHandle<'_, Key, Value> {
-        self.internal_data().get_mut(&key).unwrap()
-    }
-
-    /// Borrow a value, given a key. Returns an appropriate handle which
-    /// implements `Deref` to the value.
-    fn borrow(&self, key: Key) -> PartialStoreBorrowHandle<'_, Key, Value> {
-        self.internal_data().get(&key).unwrap()
-    }
-
-    /// The number of entries in the store.
-    fn len(&self) -> usize {
-        self.internal_data().len()
-    }
-
-    /// Whether the store is empty.
-    fn is_empty(&self) -> bool {
-        self.internal_data().is_empty()
-    }
-
-    /// Clear the store of all key-value pairs.
-    fn clear(&self) {
-        self.internal_data().clear()
-    }
-}
-
-/// Extra methods for [`PartialStore`] that require the `Value` to be `Clone`.
-pub trait PartialCloneStore<Key: Copy + Eq + Hash, Value: Clone>: PartialStore<Key, Value> {
-    /// Get a value by its key, if it exists.
-    fn get(&self, key: Key) -> Option<Value> {
-        self.internal_data().get(&key).map(|x| x.value().clone())
-    }
-
-    /// Get a value by a key, and map it to another value given its reference,
-    /// if it exists.
-    ///
-    /// It is safe to provide a closure `f` to this function that modifies the
-    /// store in some way (`create` etc). If you do not need to modify the
-    /// store, consider using [`Self::map_fast()`] instead.
-    fn map<T>(&self, key: Key, f: impl FnOnce(Option<&Value>) -> T) -> T {
-        let value = self.get(key);
-        f(value.as_ref())
-    }
-
-    /// Modify a value by a key, possibly returning another value, if it exists.
-    ///
-    /// It is safe to provide a closure `f` to this function that modifies the
-    /// store in some way (`create` etc). If you do not need to modify the
-    /// store, consider using [`Self::modify_fast()`] instead.
-    fn modify<T>(&self, key: Key, f: impl FnOnce(Option<&mut Value>) -> T) -> T {
-        let mut value = self.get(key);
-        let ret = f(value.as_mut());
-        if let Some(value) = value {
-            self.insert(key, value);
-        }
-        ret
-    }
-
-    /// Duplicate a value in the store, returning the new key.
-    ///
-    /// Assumes that the key already exists in the store.
-    fn duplicate(&self, from: Key, to: Key) {
-        if let Some(value) = self.get(from) {
-            self.insert(to, value);
-        }
-    }
-}
-
-impl<Key: Copy + Eq + Hash, Value: Clone, T: PartialStore<Key, Value>> PartialCloneStore<Key, Value>
-    for T
-{
-}
-
-/// A default implementation of [`PartialStore`].
-#[derive(Debug)]
-pub struct DefaultPartialStore<K: Eq + Hash, V> {
-    data: PartialStoreInternalData<K, V>,
-}
-
-impl<K: Eq + Hash, V> Default for DefaultPartialStore<K, V> {
-    fn default() -> Self {
-        Self { data: DashMap::with_hasher(FxBuildHasher::default()) }
-    }
-}
-
-impl<K: Eq + Hash, V> DefaultPartialStore<K, V> {
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
-
-impl<K: Copy + Eq + Hash, V> PartialStore<K, V> for DefaultPartialStore<K, V> {
-    fn internal_data(&self) -> &PartialStoreInternalData<K, V> {
-        &self.data
-    }
-}
 
 #[cfg(test)]
 mod test_super {
     use super::*;
+
     // Ensuring macros expand correctly:
-    new_store_key!(pub TestK);
     new_sequence_store_key_direct!(pub TestSeqK, TestKK);
 }
